+++ src/main/java/com/marmitex/service/TipoComidaService.java
@@ -1,10 +1,7 @@
 package com.marmitex.service;
 
 import com.marmitex.framework.CrudService;
-import com.marmitex.model.Pessoa;
 import com.marmitex.model.TipoComida;
 
-public interface TipoComidaService extends CrudService<TipoComida, Long> {
-
-
+public interface TipoComidaService extends CrudService<TipoComida, Integer> {
 }
